move prepend_a_an to base_util
