+++ droidlet/perception/semantic_parsing/nsp_templated_data_generation/template_objects/dig.py
@@ -3,10 +3,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 
 This file contains template objects associated with Dig action.
 """
-import random
-
-from ..generate_utils import *
-from ..tree_components import *
+from droidlet.base_util import prepend_a_an
 from .template_object import *
 
 #####################
